Enable coverage on a per-request basis, clean some warnings
At times it is useful to be able to enable coverage reporting
on a per-request/invocation basis. For instance, when
debugging a production system, generating a coverage report
only for invocations related to the reproduction of a
particular bug could be quite useful. Moreover, when multiple
services are involved, it's helpful to be able to propagate
the directive to start coverage to upstream services.

This PR adds the ability to "force" coverage, regardless of
pre-configured sampling ratios etc. The mechanism for doing
this is left undefined but when using a the coverband
middleware, you can pass a lambda as an option which should
return `true` to force checking on this request or `false` to
to run with existing options. That is, you can force checking
"on" but not "off".

This PR also cleans up a few warnings.